[FEATURE] Allow usage of a personal GitHub oAuth token
A new environment variable `COMPOSER_GITHUB_OAUTH_TOKEN` can
now be used to let composer authenticate against the GitHub API on
behalf of the user. This allows the builder VM to always use the
`dist` download, which in turn needs a checksum from the API
to verify downloaded zip files and to prevent fall back to the
`source` method.

See https://getcomposer.org/doc/articles/troubleshooting.md#api-rate-limit-and-oauth-tokens