Changed the name of the PHP debian packages.
Before this PR, the name of the PHP debian packages are like:

gcp-php-5.6.30
gcp-php-7.0.15

and they provide virtual packages like gcp-php56, gcp-php70.

It turned out it is problematic when we have multiple patch versions
in our repo.

For example, if we have these two packages:

gcp-php-7.0.14_7.0.14-2
gcp-php-7.0.15_7.0.15-1

in a single repo, the following command fails:

```
$ apt-get -y install gcp-php70
```

because when there are multiple candidates for a virtual package, apt
can't decide which to pick, so it fails.

So we need to change the package names as follows:

gcp-php56_5.6.30-1
gcp-php70_7.0.15-1

so that our repo will have only one package for each minor versions.