Add ability to specify which extensions you wish to build.
The parameter after the comma separated list of php versions can be a
comma separated list of extension names you wish to build.

Also, we copy the .deb packages out of the working directory into the
artifact directory instead of working in the artifact directory. We also
will not rebuild the php package if it already exists in the artifact
directory.

Fix README