Make local development easier
When doing local development and attempting to replicate the live environment, the ability to use local volumes replicated into the docker container are crucial. Unfortunately this file will then lock down the files inside the development environment as well as the files in the docker container. This adds an environment variable to turn off the lock down in the case of attempting to do the above.